+++ tests/TestCase/View/Cell/PagesWidgetsCellTest.php
@@ -148,11 +148,9 @@ class PagesWidgetsCellTest extends TestCase
         $this->assertHtml($expected, $result);
 
         //Empty on categories index
-        $request = new Request;
-        $request->env('REQUEST_URI', Router::url(['_name' => 'pagesCategories']));
-        $this->Widget = new WidgetHelper(new View($request));
-        $result = $this->Widget->widget($widget)->render();
-        $this->assertEmpty($result);
+        $widget = $this->Widget->widget($widget);
+        $widget->request->env('REQUEST_URI', Router::url(['_name' => 'pagesCategories']));
+        $this->assertEmpty($widget->render());
 
         //Tests cache
         $fromCache = Cache::read('widget_categories', $this->Pages->cache);
@@ -161,6 +159,20 @@ class PagesWidgetsCellTest extends TestCase
             'first-page-category',
             'sub-sub-page-category',
         ], array_keys($fromCache->toArray()));
+    }
+
+    /**
+     * Test for `categories()` method, with no pages
+     * @test
+     */
+    public function testCategoriesNoPages()
+    {
+        $widget = ME_CMS . '.Pages::categories';
+
+        $this->Pages->deleteAll(['id >=' => 1]);
+
+        $this->assertEmpty($this->Widget->widget($widget)->render());
+        $this->assertEmpty($this->Widget->widget($widget, ['render' => 'list'])->render());
     }
 
     /**
@@ -203,14 +215,23 @@ class PagesWidgetsCellTest extends TestCase
         $this->assertHtml($expected, $result);
 
         //Empty on categories index
-        $request = new Request;
-        $request->env('REQUEST_URI', Router::url(['_name' => 'pagesCategories']));
-        $this->Widget = new WidgetHelper(new View($request));
-        $result = $this->Widget->widget($widget)->render();
-        $this->assertEmpty($result);
+        $widget = $this->Widget->widget($widget);
+        $widget->request->env('REQUEST_URI', Router::url(['_name' => 'pagesCategories']));
+        $this->assertEmpty($widget->render());
 
         //Tests cache
         $fromCache = Cache::read('widget_list', $this->Pages->cache);
         $this->assertEquals(2, $fromCache->count());
+    }
+
+    /**
+     * Test for `pages()` method, with no pages
+     * @test
+     */
+    public function testPagesNoPages()
+    {
+        $this->Pages->deleteAll(['id >=' => 1]);
+
+        $this->assertEmpty($this->Widget->widget(ME_CMS . '.Pages::pages')->render());
     }
 }