+++ tests/TestCase/View/Cell/PostsWidgetsCellTest.php
@@ -148,11 +148,9 @@ class PostsWidgetsCellTest extends TestCase
         $this->assertHtml($expected, $result);
 
         //Empty on categories index
-        $request = new Request;
-        $request->env('REQUEST_URI', Router::url(['_name' => 'postsCategories']));
-        $this->Widget = new WidgetHelper(new View($request));
-        $result = $this->Widget->widget($widget)->render();
-        $this->assertEmpty($result);
+        $widget = $this->Widget->widget($widget);
+        $widget->request->env('REQUEST_URI', Router::url(['_name' => 'postsCategories']));
+        $this->assertEmpty($widget->render());
 
         //Tests cache
         $fromCache = Cache::read('widget_categories', $this->Posts->cache);
@@ -164,14 +162,28 @@ class PostsWidgetsCellTest extends TestCase
     }
 
     /**
+     * Test for `categories()` method, with no posts
+     * @test
+     */
+    public function testCategoriesNoPosts()
+    {
+        $widget = ME_CMS . '.Posts::categories';
+
+        $this->Posts->deleteAll(['id >=' => 1]);
+
+        $this->assertEmpty($this->Widget->widget($widget)->render());
+        $this->assertEmpty($this->Widget->widget($widget, ['render' => 'list'])->render());
+    }
+
+    /**
      * Test for `latest()` method
      * @test
      */
     public function testLatest()
     {
-        $latestPost = $this->Posts->find('active')->order(['created' => 'DESC'])->first();
-
         $widget = ME_CMS . '.Posts::latest';
+
+        $latestPost = $this->Posts->find('active')->order(['created' => 'DESC'])->first();
 
         //Tries with a limit of 1
         $result = $this->Widget->widget($widget, ['limit' => 1])->render();
@@ -235,11 +247,9 @@ class PostsWidgetsCellTest extends TestCase
         $this->assertHtml($expected, $result);
 
         //Empty on posts index
-        $request = new Request;
-        $request->env('REQUEST_URI', Router::url(['_name' => 'posts']));
-        $this->Widget = new WidgetHelper(new View($request));
-        $result = $this->Widget->widget($widget)->render();
-        $this->assertEmpty($result);
+        $widget = $this->Widget->widget($widget);
+        $widget->request->env('REQUEST_URI', Router::url(['_name' => 'posts']));
+        $this->assertEmpty($widget->render());
 
         //Tests cache
         $fromCache = Cache::read('widget_latest_1', $this->Posts->cache);
@@ -247,6 +257,17 @@ class PostsWidgetsCellTest extends TestCase
 
         $fromCache = Cache::read('widget_latest_2', $this->Posts->cache);
         $this->assertEquals(2, $fromCache->count());
+    }
+
+    /**
+     * Test for `latest()` method, with no posts
+     * @test
+     */
+    public function testLatestNoPosts()
+    {
+        $this->Posts->deleteAll(['id >=' => 1]);
+
+        $this->assertEmpty($this->Widget->widget(ME_CMS . '.Posts::latest')->render());
     }
 
     /**
@@ -317,11 +338,9 @@ class PostsWidgetsCellTest extends TestCase
         $this->assertHtml($expected, $result);
 
         //Empty on posts index
-        $request = new Request;
-        $request->env('REQUEST_URI', Router::url(['_name' => 'posts']));
-        $this->Widget = new WidgetHelper(new View($request));
-        $result = $this->Widget->widget($widget)->render();
-        $this->assertEmpty($result);
+        $widget = $this->Widget->widget($widget);
+        $widget->request->env('REQUEST_URI', Router::url(['_name' => 'posts']));
+        $this->assertEmpty($widget->render());
 
         //Tests cache
         $fromCache = Cache::read('widget_months', $this->Posts->cache);
@@ -335,6 +354,20 @@ class PostsWidgetsCellTest extends TestCase
             $this->assertInstanceOf('Cake\I18n\FrozenDate', $entity->month);
             $this->assertEquals($key, $entity->month->i18nFormat('yyyy/MM'));
         }
+    }
+
+    /**
+     * Test for `months()` method, with no posts
+     * @test
+     */
+    public function testMonthsNoPosts()
+    {
+        $widget = ME_CMS . '.Posts::months';
+
+        $this->Posts->deleteAll(['id >=' => 1]);
+
+        $this->assertEmpty($this->Widget->widget($widget)->render());
+        $this->assertEmpty($this->Widget->widget($widget, ['render' => 'list'])->render());
     }
 
     /**
@@ -373,10 +406,8 @@ class PostsWidgetsCellTest extends TestCase
         $this->assertHtml($expected, $result);
 
         //Empty on search
-        $request = new Request;
-        $request->env('REQUEST_URI', Router::url(['_name' => 'postsSearch']));
-        $this->Widget = new WidgetHelper(new View($request));
-        $result = $this->Widget->widget($widget)->render();
-        $this->assertEmpty($result);
+        $widget = $this->Widget->widget($widget);
+        $widget->request->env('REQUEST_URI', Router::url(['_name' => 'postsSearch']));
+        $this->assertEmpty($widget->render());
     }
 }