+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 2.x branch
 ## 2.19 branch
+### 2.19.2
+* fixed bug in the list of active pages for each page category;
+* fixed bug for widgets: they do not show anything if there are no records.
+
 ### 2.19.1
 * added configuration for reCAPTCHA (`config/recaptcha.php`);
 * fixed bug for `upload()` methods. Now all errors are handled properly;